coursera-scraper/requirements.txt
click>=8.1.7
rich>=13.7.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
aiohttp>=3.9.0
aiofiles>=23.2.1
pathvalidate>=3.2.0
tqdm>=4.66.0